feat(bpf): pin sid_aux union ABI at 256 + widen policy_id to u32
Set SID_AUX_PLUGIN_RAW_MAX to 256 so plugin_raw is the union's layout
anchor: sid_aux_entry is a fixed 256 bytes and behavior variants can grow
within the cap without rippling the map value size or the plugin ABI.

Widen headend_entry.policy_id u16 -> u32 to remove the SR Policy id limit.
The packed struct keeps policy_id 4-byte aligned via an explicit _pad_policy
so the bpf2go-generated Go struct matches the C layout byte-for-byte (both
204 bytes); without it Go inserts alignment padding and the b6_policy
whole-struct memcpy in maps.go shifts every following field by 2 bytes.

This is a one-time plugin ABI change: plugins compiled against the 196-byte
cap must be recompiled.

+++ pkg/bgp/apply/applier_test.go
@@ -32,8 +32,8 @@ func newFakeHeadend() *fakeHeadend {
 
 // no-op SR Policy map ops so fakeHeadend satisfies the applier's dataPlane
 // interface; the SR Policy table is unit-tested separately (srpolicy_test).
-func (f *fakeHeadend) UpsertSRPolicy(uint16, []netip.Addr) error { return nil }
-func (f *fakeHeadend) DeleteSRPolicy(uint16) error               { return nil }
+func (f *fakeHeadend) UpsertSRPolicy(uint32, []netip.Addr) error { return nil }
+func (f *fakeHeadend) DeleteSRPolicy(uint32) error               { return nil }
 
 func (f *fakeHeadend) CreateHeadendV4(p string, e *bpf.HeadendEntry, _ bpf.OwnerTag) error {
 	if f.createErr != nil {
